# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Movie Theater API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token to access protected endpoints"
      }
    },
    "schemas": {
      "Genre": {
        "type": "object",
        "properties": {
          "genre_id": {
            "type": "integer"
          },
          "genre_name": {
            "type": "string"
          }
        }
      },
      "Director": {
        "type": "object",
        "properties": {
          "director_id": {
            "type": "integer"
          },
          "director_name": {
            "type": "string"
          }
        }
      },
      "Actor": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "integer"
          },
          "actor_name": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "type": "object",
        "properties": {
          "role_id": { "type": "integer" },
          "role_name": { "type": "string" }
        }
      },
      "Membership": {
        "type": "object",
        "properties": {
          "membership_id": { "type": "integer" },
          "membership_name": { "type": "string" },
          "discount_rate": { "type": "number", "format": "float" }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "account_id": { "type": "integer" },
          "full_name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "gender": { "type": "boolean", "enum": [true, false] },
          "birthday": { "type": "string", "format": "date" },
          "id_number": { "type": "string" },
          "phone_number": { "type": "string" },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "membership_type": {
            "$ref": "#/components/schemas/Membership"
          }
        }
      },
      "AccountUpdateRequest": {
        "type": "object",
        "properties": {
          "membership_id": {
            "type": "integer",
            "nullable": true
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "boolean",
            "enum": [true, false],
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "id_number": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/accounts/{account_id}": {
      "get": {
        "tags": ["Account Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Account Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Account Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Account deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/directors/{director_id}": {
      "get": {
        "tags": ["Director Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "director_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Director Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "director_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "director_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Director"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Director Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "director_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Director deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/actors/{actor_id}": {
      "get": {
        "tags": ["Actor Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Actor Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "actor_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Actor"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Actor Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Actor deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/genres/{genre_id}": {
      "get": {
        "tags": ["Genre Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "genre_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Genre Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "genre_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "genre_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Genre Controller"],
        "parameters": [
          {
            "in": "path",
            "name": "genre_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Genre deleted successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}